WIP: Diff richtext to build model changes
Up to now, we have ignored annotation tags in the DOM when building model
changes. Insertion annotations were inferred from diff region heuristics,
and the DOM could be "corrected" from the model if the heuristics failed.

This change introduces a diff algorithm on DOM annotations, so we can
reliably set the model annotations from the DOM in all cases.

TODO:
- More tests
- Handle annotation equality more reliably

Big question: Can we use the DM linearization instead of rolling our own?
We want to consider annotation ranges rather than individual characters, and
we need to linearize quickly enough to run this twice per keystroke.

Change-Id: I8fe1706f8ba36274b0e85b553cd488a4c2c042b0