WIP: Diff richtext to build model changes
Up to now, we have ignored annotation tags in the DOM when building model
changes. Insertion annotations were inferred from diff region heuristics,
and the DOM could be "corrected" from the model if the heuristics failed.

This change introduces a diff algorithm on DOM annotations, so we can
reliably set the model annotations from the DOM in all cases.

TODO:
- Improve the diffs: several cases come out "messy" as the tests show
- Handle annotation equality more reliably

Big question: Can we use the DM linearization instead of rolling our own?
This needs to run at least twice per keystroke, so we need to linearize
quickly enough.

Change-Id: I8fe1706f8ba36274b0e85b553cd488a4c2c042b0