WIP: Diff richtext to build model changes
Up to now, we have ignored annotation tags in the DOM when building model
changes. Insertion annotations were inferred from diff region heuristics,
and the DOM could be "corrected" from the model if the heuristics failed.

This change introduces a diff algorithm on DOM annotations, so we can
reliably set the model annotations from the DOM in all cases.

TODO:
- Update the ranges handling in failing tests, and look at IME test failures
- More tests
- Profile (the linearization and diff has to run twice per keystroke)
- Document the performance characteristics more clearly

Change-Id: I8fe1706f8ba36274b0e85b553cd488a4c2c042b0