Include sequences in the command help dialog
The command help dialog currently only shows keyboard shortcuts. If we
use the sequence registry, we can add the sequences to the relevant
commands. As part of this, enhance the dialog to better distinguish
between key-combinations and key-sequences.

Currently ignored in this is a way to represent "you must be at the
start of a paragraph".

Bug: T116013
Change-Id: I21fd47f5809dce36d10edc483121ecbdc12071f6