Update patch set 3
Patch Set 3: Code-Review-1

In the tests, it gets to the first commit of removing all contents, down to to the ve.dm.Surface changeInternal, and winds up with a different selection than normal browser usage generates (range[0,2] instead of range[1,1]). Presumably for a perfectly-empty document, it _should_ be generating range[0,0], as that 2 puts it past the closing {/internallist}...

I suspect this looked like it worked superfically because the test is doing better at literally deleting everything than my basic in-browser usage is. When I'm manually doing it, there's still a paragraph there, so it's never actually dealing with a completely empty document.

Patch-set: 3
Label: Code-Review=-1