Forcibly trigger a selection when using a Tool, to close inspectors
Currently, if you have an inspector open and try to use any Tool that
does a replace (all the "format" tools are good candidates) it'll do
the Tool's edit... and then emit a selection-changed event that asks
the inspector to close. But that pops staging... which undoes the
edit... but then other setTimeout'd code expects the edit to have
happened.

The simple solution seems to be to do something that'll close all open
inspectors _before_ the Tool gets to work. Because the whole thing is
caught up in a nest of setTimeouts, the easy way to do this seems to
just be to trigger a "the selection changed to whatever the current
selection is" event, and throw the actual Tool command into its own
fairly slow setTimeout.

Bug: T76717
Change-Id: I76fcd5d431b753d3e362d365aa2c0ca4e0a4b700