Close inspectors before executing a ve.ui.Tool command
Currently, if you have an inspector open and try to use any Tool that
does a replace (all the "format" tools are good candidates) it'll do
the Tool's edit... and then emit a selection-changed event that asks
the inspector to close. But that pops staging... which undoes the
edit... but then other setTimeout'd code expects the edit to have
happened.

The simple solution seems to be to do something that'll close all open
inspectors _before_ the Tool gets to work.

Bug: T76717
Change-Id: I76fcd5d431b753d3e362d365aa2c0ca4e0a4b700