Add ve.Scheduler, use it to avoid conflict in ve.ui.Tool
We have an occasional muddle of setTimeout calls which make it hard to
wait for a clean state sometimes. So, create a Scheduler utility that
handles waiting a reasonably efficient length of time before running
whatever action is required.

As a proof of concept, use it to avoid an issue that ve.Tool has with
attempting to perform certain actions while inspectors are open.

This is the bare-minimum implementation. In the future, we may want to
make it smarter so that it can track a setTimeout stack within the
action it performs, rather than just blindly polling a test function.

Bug: T117620
Bug: T76717
Change-Id: Ib1d6a702f73cd6c99fd524761115dfff27a3199d