Update patch set 6
Patch Set 6:

It's an actual failure, but it looks like it's actually in large part that some of the TODOs in the code have been accomplished since it was written, requiring changes in the test. I'll get that done.

Patch-set: 6
Reviewer: DLynch <2915@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0