Update patch set 4
Patch Set 4: Code-Review-1

Proportional scaling is controlled by ve.dm.Scalable. If this widget is going to toggle that value there should be a setter added to ve.dm.Scalable for fixedRatio.

We may also want to define in ve.dm.Scalable if it is *allowed* to be used without a fixedRatio, and show/not-show the fixed ratio toggle accordingly.

The toggle itself is not close enough to the control. I think something along the lines of the chain icon toggle button used by the GIMP would be better.

Patch-set: 4
Label: Code-Review=-1