Add ve.ui.AnnotationContextItem
Some context items appear on annotations. For them, pressing "delete"
is ambiguous, because it might mean to delete the annotation and it
might mean to delete the annotation-plus-content. So, create a new
ContextItem abstract class which can be inherited from, to enable
a "clear annotations" button in the context popup.

Implement it on Link and Language contextitems.

(Also, we currently restrict isDeletable to only be true for aliens.)

Bug: T98272
Change-Id: I3b1b553d2e148e2a50df897b8bc5b29412bdbca7