Update patch set 3
Patch Set 3:

(1 comment)

Inspectors live inside a context menu. The DesktopContext is an inline popup, but the MobileContext lives in the toolbar.

The way this should work conceptually is the DesktopContext should be concerned with positioning and sizing itself with the surface viewport (and getting scrolled into the view), then the inspector should just make itself as big as it needs within the confines on the context.

Patch-set: 3