Update patch set 3
Patch Set 3:

Getting as much as possible into the dotfiles is handy, for the integration with editors. (Admittedly, this particular exclusions case doesn't really make a huge difference, since most editor linting plugins just do it to whatever file you're editing.)

If we want the benefits of both, perhaps we could look into having part of the build process generate the dotfiles from the central config? It's potentially more fragile, but...

Patch-set: 3