Update patch set 7
Patch Set 7:

(1 comment)

Proportional scaling is controlled by ve.dm.Scalable. If this widget is going to toggle that value there should be a setter added to ve.dm.Scalable for fixedRatio.

 -> See my comment :]

We may also want to define in ve.dm.Scalable if it is *allowed* to be used without a fixedRatio, and show/not-show the fixed ratio toggle accordingly.

 -> What should be the default? True?

The toggle itself is not close enough to the control. I think something along the lines of the chain icon toggle button used by the GIMP would be better.

 -> is that better now? https://phabricator.wikimedia.org/F3172428

Patch-set: 7
Reviewer: Florianschmidtwelzow <890@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0