Update patch set 15
Patch Set 15:

So the logic is written to handle any possible case. We can add a layer on top to prevent moves we don't think are sensible, but to begin with we should have an algorithm that tries to make a consistent table in all cases.

As such the first case is a todo for later. The second bug seems to be a valid edge case, as I don't think I considered merged cell conflicts that result in the complete destruction of a cell.

Patch-set: 15
Reviewer: Esanders <658@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0