Update patch set 14
Patch Set 14: Code-Review-1

(3 comments)

I have a few petty complaints about missing documentation, but otherwise it looks good.

As mentioned in IRC, there are some glitches around alien tables, but those are buggy anyway so it's not sufficient reason to block the rest of this. I'll open an issue about those so we can look at it later.

Patch-set: 14
Label: Code-Review=-1