FocusableNode: watch for image loaded/layout changes for invisible icons
jQuery width / height is a bit rendering-dependent, and will return 0 if
called particularly early. This results in some cases like the Math extension
appearing to not have rendering, because they just contain images. So, use
ve.Scheduler to keep an eye on not-fully-there images and make sure they're
properly set up before we initialize the icon.

Also, expand the documentation for hasRendering a bit, so it's clearer what we
consider to be "visible".

Bug: T125767
Change-Id: If322b2cde668e4ca774355fd7c02152f05ce67fa