Don't just rely on jQuery width / height for hasRendering
jQuery width / height is a bit rendering-dependent, and will return 0 if
called particularly early. This results in some cases like the Math extension
appearing to not have rendering, because they just contain images. However,
the raw DOM attribute contains a correct value, so we can check that as well.

Expand the documentation for hasRendering a bit, so it's clearer what we
consider to be "visible".

Bug: T125767
Change-Id: If322b2cde668e4ca774355fd7c02152f05ce67fa