LinkContextItem: add a button to select the entire label
Adds a button which quickly selects the entire contents of the link
annotation. This makes it far easier to select the entire label for editing,
without accidentally selecting outside the annotation and thus removing it.

Refactor a bit of the AnnotationContextItem so that the "do something to all
related annotations" code is reusable.

Bug: T124305
Change-Id: Iad48fb559f3dae16cc0013791b79ec71308a1f9f