Update patch set 1
Patch Set 1:

The underlying cause is *probably* that the preview dialog occasionally destroys and replaces its MWMathNode (which sets the model to null), but this doesn't cancel any outstanding promises that are part of the async content generation attached to that node. The timing on persuading this to happen is a bit fiddly, so it's hard to reliably reproduce it.

The change here is basically an improvement to the existing check in this class for "have I been discarded?" so it handles "have I been destroyed?" as well.

Patch-set: 1
Reviewer: DLynch <2915@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0