FindAndReplace: Avoid getNativeRange() optimisation
When the dialog is opened the view and model are momentarily
out of sync, in that the native range has been removed, but the
model selection is not yet null. If you open the dialog with a
selection, the subsequent call to get rects for that range will
trigger a optimisation in getNativeRange to use the current
native range, which is now in the dialog, not the surface.

Change-Id: Ic9a7475cbd8a22cbc7b277c081ac40f1a21b859e