Update patch set 3
Patch Set 3:

An "edit label" inspector is totally possible.

It has some possible drawbacks, though. The simple case of editing a text label is easily served. It's more of a pain to handle everything that could possibly be inside a link (images, styled text, etc), while also remembering all the things which /can't/ be in a link -- that was the appeal of just setting the selection, not having to even think about all that.

We do already handle much of that case in areas like image captions, apart from the "stuff that's not allowed in a link" bit, so it's certainly doable that way. It feels less elegant than managing to fall back to the existing document rendering, though.

Patch-set: 3
Label: Verified=0