LinkAnnotationInspector: add a button to edit the label
Adds a button which quickly selects the entire contents of the link
annotation. This makes it far easier to select the entire label for editing,
without accidentally selecting outside the annotation and thus removing it.

This exactly mirrors the behavior of the mediawiki linknodeinspector, when
converting from a numbered link to a labeled one.

Bug: T124305
Change-Id: I5f36835fefaecd238ae66795a9c3abf348359c60