Update patch set 1
Patch Set 1:

Hmm. On typing, this replaces the selection with an unannotated character for me (both Chrome and Firefox), whereas double-clicking to select the word and typing does not.

Also, we'd have to come up with a design for where to put this button (and the "add label" button of the very-rare MWLinkNodeInspector, hence why we've not fixed it yet), but this has promise.

Patch-set: 1
Label: Verified=0