Update patch set 10
Patch Set 10:

None of this is needed aside from a basic entry point (see sync-gh-pages.sh). We already have all this. We have a working demos directory. https://wikimedia.github.io/VisualEditor/

The only thing that needs to happen is inside a Jenkins job (not inside this repo) to run the build script and publish the demos directory. The one we already have. Can be combined with the coverage publisher probably, similarly to what we do for oojs-ui.

Patch-set: 10
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0