Create index.html in bin folder
This is so we can copy it with grunt when we are generating the demos at
doc.wikimedia.org

This is partially a copy of sync-gh-pages.sh but instead of generating it
with a script that would be duplicat we can just copy it.

Also add a new grunt task publish-build which will build the stuff we need
for demos.

Also make prepublish -> predoc since prepublish is basically install or
preinstall making it run duplicat tests.

Change-Id: I028de26e5eee38987f669e28b16d69434f382f41