Update patch set 4
Patch Set 4: Code-Review-1

So by the spec we shouldn't strip empty lines. An empty line in the middle of a CSV could be meaningful. If it is common that CSV generators are adding trailing line breaks we could ignore those, but that should be limited to the last line maybe?

If we are going to strip all empty lines we should just use line.trim() === ''

Patch-set: 4
Label: Code-Review=-1