Update patch set 1
Patch Set 1:

Needs David C's input. This works by observing a bad selection and fixing it because the "activeNode" property is in the view. If you double click and type *really* quickly it is possible to beat the setTimeout and overwrite the activeNode boundary, although we could fix that separately.

Patch-set: 1
Label: Verified=0