Introduce dm.Surface#getModifiedRanges
Computes a minimal set of ranges covering all modifications
to the surface. This would useful for highlighting changes.

As it rebuilds the list from scratch every time it has N^2 complexity
with regards to the transaction list length, so for showing this
in real time it might need some caching (cf SurfaceFragments).

Change-Id: I973e9794cc0c0dfeadede6e5414fb2cf5eaf7847