Update patch set 1
Patch Set 1:

Sort of. It's for situations where you want to break the association between the tabable elements and the rest of the page entirely -- my imagined use case is a modal dialog, where having the focus escape to something outside the modal is deeply user-unfriendly. (Particularly since we have a default skin which uses tabindex on the site logo...)

Granted, in such cases you'd probably need to be particularly zealous about making sure all plausible tab targets were registered, to avoid in-modal accessibility issues.

Patch-set: 1