Update patch set 1
Patch Set 1:

Not particularly -- OO.ui.Window sets up what it calls focus traps and inserts then in the DOM immediately before / after a Window. It uses them to make the focus loop inside the Window, with the same missing-the-location-bar issues as this code has.

That said, its existence probably does mean that the loop / trap code can be stripped out, since I imagine that anywhere we'd want that behavior we'll be using an OO.ui.Window, and we can fall back on it. I can also add a little more code in to make sure that this works properly with that -- as-is, this code wouldn't *force* something to be focusable as the trap would consider it.

Patch-set: 1
Label: Verified=0