Insert undo breakpoints at word breaks
When users interact with the contenteditable and enter text, check the last
few characters and see if we have a word-characters followed by a non-word-
character. If so, set a breakpoint. For languages with word breaks, this is a
fairly accurate heuristic for having undo work per-word.

Leave the existing interval behavior in place, with a much longer interval, as
a fallback for people working in languages with infrequent non-word
characters, or people who're entering very long words slowly.

In languages with word breaks

Bug: T127834
Change-Id: I937df66706e533687f327396b76f75d686cfacd1