Insert undo breakpoints at word breaks
When users interact with the contenteditable and enter text, check if we seem
to have caused a word break. If so, set a history breakpoint.

Leave the existing interval behavior in place, with a much longer interval, as
a fallback for people working in languages with infrequent non-word
characters, or people who're entering very long words slowly.

Bug: T127834
Change-Id: I937df66706e533687f327396b76f75d686cfacd1