Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Looks promising.

A complication is that ve.ce.getOffset needs to work when the DOM has been modified by native ContentEditable actions (e.g. by typing text), before the model has been updated. It has to be before the model update, because the data is used to build a transaction.  An example is as follows:

1. Edit <p><b><i>a<!-- foo -->b</i></b></p> 

2. Put cursor just before 'a' (at { node: #text('a'), offset: 0 } )

3. Press 'x' and you get the following call stack (abbreviated):

ve.ce.Surface.afterDocumentKeyDown
-> ve.ce.SurfaceObserver.pollOnceInternal
   -> ve.ce.RangeState.saveState
      -> ve.ce.veRangeFromSelection
         -> ve.ce.getOffset
	    -> countInvisibleNodes

At the time of the call to countInvisibleNodes, the DOM has the inserted 'x' but the model does not, so it thinks offset 1 inside the paragraph is *after* the invisible comment nodes, when in fact it should be before.

Theoretically we could try to infer what text has been modified based on the selection, but in practice it can be tricky to do so in an IME-safe way, because IMEs can be quirky about how much text is selected.

So perhaps we ought to stop creating ve.Range objects in ve.ce.RangeState entirely, and instead calculate them in ve.ce.Surface.handleObservedChanges where we can better evaluate the changes against the model.

Patch-set: 1
Reviewer: Divec <839@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1