Update patch set 2
Abandoned

This fixes some cases, but makes others worse. Investigation with David Chan confirmed that this is arguably a WebKit browser bug (with Firefox having a similar-but-different one), and browser tickets have been created.

If we assume browsers won't fix it, we could perhaps approach this in the linear key handler, checking whether the selection hasn't changed when we think it should. (This is, of course, fragile. "Should the selection have changed?" is a complicated question, and we abdicate it to the browsers for a reason.)

Patch-set: 2
Status: abandoned