Update patch set 9
Patch Set 9:

> Do you really think a tool for adding <big> is a good idea?

I think that it's necessary for now to provide feature-parity with the wikitext editor in VE-MW.

I'm in favour of deprecating them from wikitext in favour of something different, though the "obvious" answer of <span style="font-size: larger;">s violates the plan to remove inline styling options.

Once we discourage it from wikitext we'd obviously switch these over to use that in preference, but we're not likely to see that soon.

Patch-set: 9
Label: Verified=0