LinkAnnotationInspector: add a "label" field
This field will allow the link text to be edited inside the inspector. If left
empty, the current behavior of generating link text from the href will be
followed. If the current value of the label contains anything which isn't
text, the input will be disabled and its value ignored on teardown.

Bug: T55973
Change-Id: Ic5fbbf3629e92d933fda61969615ff5262437e07