Update patch set 1
Patch Set 1: Code-Review-2

This creates a TextInputWidget which is just wrong; the contents covered by the link annotation could be anything up to and including <span>s, <img>s and so on; this would need to be a complete SurfaceWidget, which would totally dominate this inspector and destroy the whole design of it being a lightweight, simple control. I'm pretty strongly against this change for exactly this reason.

Patch-set: 1
Reviewer: Jforrester <366@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-2