Update patch set 1
Patch Set 1:

James: This is totally cribbing from Google Docs' design for link labels. It shows the field... if the label is simple. If it's not simple, it hides it. That lets us give people a field that works in most cases, and won't break anything otherwise.

Admittedly, I just have it *disabled* if it's not-simple, on the theory that might be less confusing, but I could go full-hidden.

Patch-set: 1
Label: Verified=0