Update patch set 5
Patch Set 5:

On investigation, there aren't any select-all issues, and applying this actually gets rid of a drag-to-select issue in Firefox (currently dragging over a slug loses your selection before the slug).

However! Applying this patch causes some difficulties moving the selection by cursoring in Firefox. Specifically, if you try to cursor *up* past a slug, it'll jump to the start of the document. Cursoring up in Chrome will sometimes skip it if you're coming out of a list.

The patch does maintain keyboard-focus for the node, if it's not skipped, and focusing it and hitting enter to trigger a new paragraph still works.

Patch-set: 5
Reviewer: DLynch <2915@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0