Update patch set 2
Patch Set 2:

That is technically true, however in almost all cases additional annotations to the link are covering, and can be preserved. This is the same behaviour as Google Docs. Links containing content (e.g. img) are many times rarer still. For these cases we can block/and or warn (or not).

Patch-set: 2