WIP/EXPERIMENTAL Tree differ to run on dm.DocumentNode
reconcile.js is very fast, but relies on stable IDs, which are
not always available, so it often won't find the minimal diff.
This isn't ideal for highlighting differences. The tree differ
finds the minimal diff, but currently with a heavy performance
cost, as it treats the entire documents as single trees and
searches all possible alignments.

Given the predictable properties of real-world document changes,
it may be possible to use some simple heuristics to vastly reduce
the search space for the tree differ and still find the minimal
diff in the majority of cases. For example, edits do not tend to
affect many different children of the dm.DocumentNode.

Next steps: (1) treat the documents as multiple trees, rooted at
the children of the document node. (2) Exclude identical trees
before doing the full diff.

Change-Id: I30162f5d24e53de84455b62f19fd0bb3c7069489