WIP/EXPERIMENTAL Make ve.dm.VisualDiff
A ve.dm.VisualDiff gets everything that is needed to build a
visualisation of the diff:
- a mapping of old nodes to their analogous new nodes (changed
or unchanged)
- the moves that map a changed old node to its analogous new
node
- a list of removed nodes
- a list of inserted nodes
NB Nodes are ordered, so it is also possible to infer moves.

Next steps: build something basic for visualising the diff in
VE UI - to be adapted according to collaboration with design.

Change-Id: I23295efd75e9e37b273b99a361ded99a98217295