Update patch set 14
Patch Set 14:

This shouldn't really be called "rebasing", it should be called "transpose".  The fundamental property needed is that given operations A and B which apply to document D, it returns a pair of operations A' and B' such that B'(A(D)) = A'(B(D)).

It is not required that the result "preserve" changes made by A or B, necessarily.  You can implement the transpose by (deterministically) selecting one of A or B to throw away, such that (say) A'(d) = A(!B(d)) and B' = noop (where !B means "the inverse of B" aka "the operation which undoes the effect of B").

See https://gerrit.wikimedia.org/r/#/c/151545/8/src/dm/ve.dm.Transaction.transpose.js for more discussion.  In my experience you need a pretty hefty test suite to get confidence in these operations, since there are lots of weird corner cases.

Patch-set: 14
Reviewer: C. Scott Ananian <664@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0