ve.init.Target: Wrap floating toolbar init in rAF
Determining whether we need to float or unfloat the toolbar
requires us to measure the position of the toolbar and the
scrollTop of its container, which causes a slow forced reflow.

Even with rAF this is still kind of slow, and we should probably
defer it more, because ve.ui.Surface#initialize still has
to wait for it.

Change-Id: Id189484e77c5dd6f676a5524fcb616c5e9ac7ea4