WIP/EXPERIMENTAL Make ve.dm.VisualDiff and ve.ui.DiffElement
A ve.dm.VisualDiff gets everything that is needed to build a
visualisation of the diff:
- a mapping of old nodes to their analogous new nodes (changed
or unchanged)
- the moves that map a changed old node to its analogous new
node
- a list of removed nodes
- a list of inserted nodes
NB Nodes are ordered, so it is also possible to infer moves.

A ve.ui.DiffElement displays the diff. Currently assumes a
unified diff but could also produce a side-by-side diff.

Change-Id: I23295efd75e9e37b273b99a361ded99a98217295