Update patch set 1
Patch Set 1:

I'm not sure there are no other cases where we might have a deactivated selection then change it. Or we may have such cases in the future. Is there a way to restrict this to the case described in the bug?

Patch-set: 1
Reviewer: Esanders <658@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0