Follow-up e3192e2d3: property fix layout thrashing in updateInvisibleIcon
Rather than deferring the whole function with rAF, do the measurements
immediately and defer the DOM modification. In theory, this means we
should be able to do all the measurements first and all the DOM modifications
after that, causing no reflows. In practice, there are still a couple
of reflows because of other stuff that happens during FocusableNode
setup, so we'd need to use something like the fastdom library to
really cut down on reflows.

This cuts post-AJAX load time on [[en:Barack Obama]] down to 5.8s
down from 7.6s on my laptop.

Change-Id: I4aff88f705782d7ab3bd6f7aaa8096324a29b813