[WIP] Remove queueing of incoming changes
When a change comes in while we have a change in flight,
don't queue them up to apply them later, but apply them
immediately. This requires rebasing the submitted change
(and any unsubmitted local changes) over each remote change
as it comes in, and it means that when our submitted change
does arrive, we only have to update some state.

We also need to handle conflicts differently. When a change
arrives that conflicts with our local changes, we first try
to rebase it again, but only over our submitted change.
If that succeeds, we discard our unsubmitted local changes
and keep the submitted change, expecting it to be confirmed
(or rejected) by the server later. If the smaller rebase
also conflicts, we discard all local changes.

Finally, we need to use the transactionStart field from the
rejectedUpdate event to figure out which change conflicted,
and ignore events telling us about conflicts that we've already
detected locally and dealt with. I strongly suspect that this
is always the case, and that conflict notifications from the
server are now redundant.

Change-Id: I8bb34c1122558241ad0f43bfd96e7e6b1805a90d