[WIP] Remove queueing of incoming changes
When a change comes in while we have a change in flight,
don't queue them up to apply them later, but apply them
immediately. This requires rebasing the submitted change
(and any unsubmitted local changes) over each remote change
as it comes in, and it means that when our submitted change
does arrive, we only have to update some state.

TODO: I probably need to rewrite history
TODO: Test undo on conflict
TODO: Try to support multiple changes in flight

Change-Id: I8bb34c1122558241ad0f43bfd96e7e6b1805a90d