[VERY WIP] Change continuation
The first part of my approach to having multiple changes in flight.
This code allows clients to submit changes based on top of previously
submitted changes before they've seen a confirmation/rejection
by rebasing them as they come in and storing the result of the previous rebase.

Still needs integration, testing, and a proof by induction for
the iterative part (i.e. the reuse of the rebase result); the base case
is easier to prove, that proof is in the comments already.

Change-Id: I76c77a74ee750c75e2aa218081a3956045f88485