Converter: Remove internal during the main loop
It was placed in a separate loop so as not to affect whitespace
calculations at the end of the document, so fix that calculation
to include the next node being the internal list.

Also assume the internalList is the last thing in the document.
We were over cautious about this when IL was introdued, but we
make this assumption is so many other places.

Change-Id: Ice6b16722b234b63548be72a8e1017ef24d3cc8f