Update patch set 8
Patch Set 8:

@DLynch: I've rebased onto the test harness improvements.

As things stand we can still drag so that the selection anchor is outside the active node and the selection focus is inside.

Also, in Firefox if you have a block image caption 'abc' then you can't click to move the selection into the interior of the text node ('a|bc' or 'ab|c'). I can't work out why; there may be an underlying ce=false quirk in Firefox.

Patch-set: 8
Label: Verified=0