Update patch set 8
Patch Set 8:

I tested a simple true/false/true CE nested in Firefox using CE-sandbox and it seems to capture the cursor correctly - so must be something we are doing. Also in FF I can't cursor up/down inside a caption node (the cursor jumps back to the start of the caption).

Patch-set: 8
Label: Verified=0