Update patch set 1
Patch Set 1:

If the -1 is for the tools not working when HTML is selected, I don't intend to fix that yet. The purpose of this commit is to move more logic upstream to have a basic framework for non-MW source mode.

Making an actually useful HTML source editor is not as high priority.

Patch-set: 1
Label: Verified=0