Update patch set 2
Patch Set 2: -Code-Review

> If the -1 is for the tools not working when HTML is selected, I don't intend to fix that yet. The purpose of this commit is to move more logic upstream to have a basic framework for non-MW source mode.

My main concern is checking if this means we've got the model wrong -- presumably this is breaking because the tools are treating the content as plain text strings and HTML escaping it, and if that's happening at the "wrong" level can we ever actually make this work?

If you're OK with it I'll +2.

Patch-set: 2