Update patch set 2
Patch Set 2:

Yes - we have the same bug in MW. If you select '''foo''' and make it a link you get nowikis. This is a known issue. In most cases it's fixed by writing custom tools, e.g. replace 'make bold' with '(un)wrap "<b>"'.

A generic fix would be to parse the selected text into DM before throwing it in the dummy surface. That's easy for HTML, but for wikitext it's another round-trip. We could also take the approach of replacing the selected text with a unicode placeholder, then dropping the original text back in wherever that placeholder ends up.

Patch-set: 2
Reviewer: Esanders <658@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0