Update patch set 1
Patch Set 1:

This is probably correct but it means we have to come up with a new tool for inserting block quote, as it is currently treated as a CBN conversion in the format dropdown. I think we did it that way because most blockquotes are only one paragraph.

Patch-set: 1