Update patch set 2
Patch Set 2:

It's used in TableSelection to expand based on the surrounding table node. We can probably still remove it, making it an argument - but that should be done later. In the meantime this is breaking table collaboration.

Patch-set: 2
Reviewer: Esanders <658@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0