Update patch set 1
Patch Set 1:

To explain the issue in more detail, this is two things happening:

In source mode, the annotation inspector doesn't have access to the previousSelection property, which it assumes exists when the window closes. This causes the endless-loop of it trying to close, having an error and staying open, etc. Patch 330439 stops that assumption from being made.

When that's fixed, there's still nothing being inserted if you started with a collapsed selection. That's because NWE's annotation method doesn't cope well with empty annotations (they get boiled down to "" in the wikitext-generation, as far as I can tell). As such, *this* patch provides a fallback in this case, mirroring the behavior of the italic/etc commands (which generate markup like `''Italic''` in the collapsed selection case).

Patch-set: 1
Reviewer: DLynch <2915@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0