Allow custom paste string types to be set
The string types which will be considered for paste are hardcoded, making it
impossible to allow a new string handler to be registered for paste events.
The order is important, so just falling back to the `kind` property would have
side-effects.

This allows the init Target to override the paste string types, presumably if
the implementor has registered a dataTransferHandler for a custom mime type.

Change-Id: Ibd458859b01de982487fefeabb901cbb61996800