DocumentSynchronizer: Simplify pushRebuild() API
Instead of asking for an old range and a new range, ask
for a range (the old range) and a length change. This is
equivalent because the new range can be computed from the
old range, the adjustment, and the length change, which is
exactly how TransactionProcessor computed this argument anyway.

Change-Id: Ibca625ccd05add8d77e535ed8e3d0b1fb6ab5d77