Honorably discharge ve.dm.DocumentSynchronizer
For almost five years, DocumentSynchronizer has served us well,
unfailingly synchronizing (valid) linear model transactions to
the DM tree, and only making me buy the entire team lunch once[1].

However, David Chan's metadata changes ran into a bug in
DocumentSynchronizer, and once we started looking into it we kept
finding more cases where processing a somewhat unusual but
completely valid transaction would cause DocumentSynchronizer to crash.
Because I don't want to go bankrupt buying everyone lunch for
all eternity, it's time to retire DocumentSynchronizer.

The failing cases we found were all to do with overlapping rebuild
ranges: inserting text into a paragraph and then merging that paragraph
will involve a resize of the text node but then a rebuild of the
paragraph containing that text node; merging p1 and p2 but then
also p2 and p3 will cause a rebuild for p1+p2 and a rebuild for
p2+p3; and so on. These cases are very hard to deal with in a world
where the entire linmod is updated first before the tree is synchronized
based on data gathered during the linmod update phase. I tried to do this
and it's probably possible to cover all the bases this way,
but it makes the code much more complex and difficult to understand.

Instead, this change makes TransactionProcessor responsible for
both linmod updates and tree sync. Every modification to the linmod
is immediately synchronized to the tree (except for unbalanced replace
operations, which are buffered until balance is achieved).
This makes the code simpler (~280 fewer lines of code) and easier to understand
and reason about, fixes all known bugs with overlapping rebuild ranges,
and also makes tree synchronization more efficient because fewer
tree traversals are needed (as this tight integration allows us to
use the output of selectNodes() more effectively).

 _
 |
 |
 |
 |^`-,._.,-'^`-,.
 |              |
 | visua|       |
 |      |editor |
 |              |
 |^`-,._.,-'^`-,.
 |
 |
 |
RIP DOCUMENTSYNCHRONIZER
2012 - 2017

[1] See dc646ff3 AKA https://gerrit.wikimedia.org/12345

Change-Id: Ia9cda655b2c0060a0103ab3fa6c74a86d504697f