Update patch set 1
Patch Set 1:

Mostly I'm unsure about how clear to users it will be to "confirm" or "cancel" a dialog. We don't use this naming internally, so it could be confusing to developers too (in OOjs UI parlance, this is called "execute primary action" and "execute safe action"). Wording suggestions?

Patch-set: 1