Update patch set 2
Patch Set 2:

I think this is nonspecific enough that people won't understand what it means as-is. "...in a popup" or similar could help, but I'm not sure. "Press the blue button if one exists" is as simplified as I can think of it, but has obvious flaws.

Command-help dialog does sort of poorly at handling those cases where an action is only performable in certain contexts, in general.

Patch-set: 2
Reviewer: DLynch <2915@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0