Update patch set 5
Patch Set 5:

(1 comment)

Yeah maybe let's document that as unsupported. It'd be good to add a failing test case (i.e. a test case that asserts that this fails).

I'm not sure if we should plan to fix this very soon, since your tree-based approach, if it takes off, would deal with this for free. If the tree thing doesn't work out though, then we do need to figure out a way to make this code handle this case, probably with a parent stack.

Patch-set: 5
Reviewer: Catrope <5@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0