Update patch set 1
Patch Set 1:

That's a good question. The code works by observing the annotation ranges in newDoc, right?

Clearly if the annotation extends strictly beyond the change range then it should not be split, and if the annotated range is strictly inside (smaller) than the change range then it should lie inside the change tags.

But if the annotation range is exactly the same as the change range, then doesn't it depend on the change? Consider Consider

  (1) [-<i>bar</i>-][+<b>bar</b>+]
  (2) <b>[-foo-][+bar+]</b>
  (3) <b>ba[-r-][+z+]</b>

Presumably, in an ideal world you would want to place the span tags as the representations above apply. But you can only distinguish (1) and (2) by looking at oldDoc.

If this is too complicated to implement right now, you could preserve just annotations that cover the selection and at least one character before or after it.

Patch-set: 1