TransactionProcessor: modify DM tree branches instead of rebuilding them
ve.dm.TransactionProcessor:
* Replace prior tree rebuilds with calls to TreeModifier

ve.dm.TreeModifier:
* Modify a DM tree branch based on a transaction

ve.dm.TreeCursor:
* Cursor used by TreeModifier

tests:
* Fix some bugs in the test data caught by the new stricter code

TODO:

1. The tree is restructured after the linear modification. It would now make more sense to
change both the linear model and the tree at the same time. Ideally this would be in a follow-up
commit, but if so we need a temporary fix for things like emitAnnotate that falsely assume the
tree matches the linear model.

Bug: T162762
Change-Id: I22d32e7f3a3afb176367f2163acb4c123e8f0ee0